split server and index
